dom/xul/test/test_bug1061864_2.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1061864
-->
<window title="Mozilla Bug 1061864"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul" onload="RunTest();">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 1061864 **/
  SimpleTest.waitForExplicitFinish();

  function RunTest()
  {
    // Test that the docshell prerendered flag will be correctly inherited in
    // prerendered documents.
    test(false, function() {
      test(true, function() {
        SimpleTest.finish();
      });
    });
  }

  function test(prerendered, callback) {
    var parentIframe = document.createElement("iframe");
    if (prerendered) {
      parentIframe.setIsPrerendered();
    }
    parentIframe.setAttribute("src", "1061864.html");
    parentIframe.onload = function() {
      var childIframe = parentIframe.contentDocument.getElementById("childiframe");
      var childDocShell = childIframe.frameLoader.docShell;
      is(childDocShell.isPrerendered, prerendered, "The docshell is" + (prerendered ? "" : " not") + " prerendered");
      callback();
    }
    document.documentElement.appendChild(parentIframe);
  }
  ]]>
  </script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1061864"
     target="_blank">Mozilla Bug 1061864</a>
  </body>
</window>
